+++ src/rust/src/wallet_scanner.rs
@@ -10,6 +10,11 @@ use std::sync::{
 
 use crossbeam_channel as channel;
 use memuse::DynamicUsage;
+use nonempty::NonEmpty;
+use orchard::{
+    note_encryption::OrchardDomain,
+    primitives::redpallas::{self, SpendAuth},
+};
 use sapling::{bundle::GrothProofBytes, note_encryption::SaplingDomain};
 use zcash_note_encryption::{batch, BatchDomain, Domain, ShieldedOutput, ENC_CIPHERTEXT_SIZE};
 use zcash_primitives::{
@@ -42,16 +47,20 @@ impl OutputDomain for SaplingDomain {
     const KIND: &'static str = "sapling";
 }
 
+impl OutputDomain for OrchardDomain {
+    const KIND: &'static str = "orchard";
+}
+
 /// A decrypted note.
-struct DecryptedNote<A, D: Domain> {
+pub(crate) struct DecryptedNote<A, D: Domain> {
     /// The tag corresponding to the incoming viewing key used to decrypt the note.
-    ivk_tag: A,
+    pub(crate) ivk_tag: A,
     /// The recipient of the note.
-    recipient: D::Recipient,
+    pub(crate) recipient: D::Recipient,
     /// The note!
-    note: D::Note,
+    pub(crate) note: D::Note,
     /// The memo sent with the note.
-    memo: D::Memo,
+    pub(crate) memo: D::Memo,
 }
 
 impl<A, D: Domain> fmt::Debug for DecryptedNote<A, D>
@@ -330,7 +339,50 @@ where
     }
 }
 
-impl<A, D: BatchDomain, Output: ShieldedOutput<D, ENC_CIPHERTEXT_SIZE> + Clone>
+/// Helper trait to abstract over how Sapling and Orchard outputs are represented.
+trait OutputSource<'o, Output: 'o> {
+    type Iterator: Iterator<Item = &'o Output>;
+
+    /// Returns the number of outputs in this source.
+    ///
+    /// This enables sources to be used that know their length but don't provide iterators
+    /// that implement `ExactSizeIterator`.
+    fn count(&self) -> usize;
+
+    /// Returns an iterator over references to the outputs.
+    fn get(&'o self) -> Self::Iterator;
+}
+
+impl<'o> OutputSource<'o, OutputDescription<GrothProofBytes>>
+    for &'o [OutputDescription<GrothProofBytes>]
+{
+    type Iterator = std::slice::Iter<'o, OutputDescription<GrothProofBytes>>;
+
+    fn count(&self) -> usize {
+        self.len()
+    }
+
+    fn get(&'o self) -> Self::Iterator {
+        self.iter()
+    }
+}
+
+impl<'o> OutputSource<'o, orchard::Action<redpallas::Signature<SpendAuth>>>
+    for NonEmpty<orchard::Action<redpallas::Signature<SpendAuth>>>
+{
+    type Iterator =
+        Box<dyn Iterator<Item = &'o orchard::Action<redpallas::Signature<SpendAuth>>> + 'o>;
+
+    fn count(&self) -> usize {
+        self.len()
+    }
+
+    fn get(&'o self) -> Self::Iterator {
+        Box::new(self.iter())
+    }
+}
+
+impl<'o, A, D: BatchDomain, Output: ShieldedOutput<D, ENC_CIPHERTEXT_SIZE> + Clone + 'o>
     Batch<A, D, Output>
 {
     /// Adds the given outputs to this batch.
@@ -338,18 +390,23 @@ impl<A, D: BatchDomain, Output: ShieldedOutput<D, ENC_CIPHERTEXT_SIZE> + Clone>
     /// `replier` will be called with the result of every output.
     fn add_outputs(
         &mut self,
-        domain: impl Fn() -> D,
-        outputs: &[Output],
+        domain: impl Fn(&Output) -> D,
+        outputs: &'o impl OutputSource<'o, Output>,
         replier: channel::Sender<OutputItem<A, D>>,
     ) {
-        self.outputs
-            .extend(outputs.iter().cloned().map(|output| (domain(), output)));
-        self.repliers.extend((0..outputs.len()).map(|output_index| {
-            OutputReplier(OutputIndex {
-                output_index,
-                value: replier.clone(),
-            })
-        }));
+        self.outputs.extend(
+            outputs
+                .get()
+                .cloned()
+                .map(|output| (domain(&output), output)),
+        );
+        self.repliers
+            .extend((0..outputs.count()).map(|output_index| {
+                OutputReplier(OutputIndex {
+                    output_index,
+                    value: replier.clone(),
+                })
+            }));
     }
 }
 
@@ -487,12 +544,12 @@ where
     /// If after adding the given outputs, the accumulated batch size is at least
     /// `BATCH_SIZE_THRESHOLD`, `Self::flush` is called. Subsequent calls to
     /// `Self::add_outputs` will be accumulated into a new batch.
-    fn add_outputs(
+    fn add_outputs<'o>(
         &mut self,
         block_tag: BlockHash,
         txid: TxId,
-        domain: impl Fn() -> D,
-        outputs: &[Output],
+        domain: impl Fn(&Output) -> D,
+        outputs: &'o impl OutputSource<'o, Output>,
     ) {
         let (tx, rx) = channel::unbounded();
         self.acc.add_outputs(domain, outputs, tx);
@@ -552,25 +609,64 @@ where
 type SaplingRunner =
     BatchRunner<[u8; 32], SaplingDomain, OutputDescription<GrothProofBytes>, WithUsage>;
 
+type OrchardRunner = BatchRunner<
+    [u8; 64],
+    OrchardDomain,
+    orchard::Action<redpallas::Signature<SpendAuth>>,
+    WithUsage,
+>;
+
 /// A batch scanner for the `zcashd` wallet.
 pub(crate) struct BatchScanner {
     params: Network,
     sapling_runner: Option<SaplingRunner>,
+    orchard_runner: Option<OrchardRunner>,
 }
 
 impl DynamicUsage for BatchScanner {
     fn dynamic_usage(&self) -> usize {
-        self.sapling_runner.dynamic_usage()
+        self.params.dynamic_usage()
+            + self.sapling_runner.dynamic_usage()
+            + self.orchard_runner.dynamic_usage()
     }
 
     fn dynamic_usage_bounds(&self) -> (usize, Option<usize>) {
-        self.sapling_runner.dynamic_usage_bounds()
+        let (params_lower, params_upper) = self.params.dynamic_usage_bounds();
+        let (sapling_lower, sapling_upper) = self.sapling_runner.dynamic_usage_bounds();
+        let (orchard_lower, orchard_upper) = self.orchard_runner.dynamic_usage_bounds();
+
+        (
+            params_lower + sapling_lower + orchard_lower,
+            params_upper
+                .zip(sapling_upper)
+                .zip(orchard_upper)
+                .map(|((a, b), c)| a + b + c),
+        )
+    }
+}
+
+pub struct OrchardPreparedIncomingViewingKeys(
+    Vec<([u8; 64], orchard::keys::PreparedIncomingViewingKey)>,
+);
+
+impl OrchardPreparedIncomingViewingKeys {
+    pub(crate) fn new<'a>(
+        ivks: impl Iterator<Item = &'a orchard::keys::IncomingViewingKey>,
+    ) -> Self {
+        Self(
+            ivks.map(|ivk| {
+                let prepared_ivk = orchard::keys::PreparedIncomingViewingKey::new(ivk);
+                (ivk.to_bytes(), prepared_ivk)
+            })
+            .collect(),
+        )
     }
 }
 
 pub(crate) fn init_batch_scanner(
     network: &Network,
     sapling_ivks: &[[u8; 32]],
+    orchard_ivks: *mut OrchardPreparedIncomingViewingKeys,
 ) -> Result<Box<BatchScanner>, &'static str> {
     let sapling_runner = if sapling_ivks.is_empty() {
         None
@@ -586,9 +682,17 @@ pub(crate) fn init_batch_scanner(
         Some(BatchRunner::new(ivks.into_iter()))
     };
 
+    let orchard_runner = if orchard_ivks.is_null() {
+        None
+    } else {
+        let ivks = unsafe { Box::from_raw(orchard_ivks) };
+        Some(BatchRunner::new(ivks.0.into_iter()))
+    };
+
     Ok(Box::new(BatchScanner {
         params: *network,
         sapling_runner,
+        orchard_runner,
     }))
 }
 
@@ -622,9 +726,15 @@ impl BatchScanner {
             runner.add_outputs(
                 block_tag,
                 txid,
-                || SaplingDomain::new(sapling_serialization::zip212_enforcement(&params, height)),
-                bundle.shielded_outputs(),
+                |_| SaplingDomain::new(sapling_serialization::zip212_enforcement(&params, height)),
+                &bundle.shielded_outputs(),
             );
+        }
+
+        // If we have any Orchard IVKs, and the transaction has any Orchard actions, queue
+        // the actions for trial decryption.
+        if let Some((runner, bundle)) = self.orchard_runner.as_mut().zip(tx.orchard_bundle()) {
+            runner.add_outputs(block_tag, txid, OrchardDomain::for_action, bundle.actions());
         }
 
         // Update the size of the batch scanner.
@@ -638,6 +748,9 @@ impl BatchScanner {
     /// Subsequent calls to `Self::add_transaction` will be accumulated into new batches.
     pub(crate) fn flush(&mut self) {
         if let Some(runner) = &mut self.sapling_runner {
+            runner.flush();
+        }
+        if let Some(runner) = &mut self.orchard_runner {
             runner.flush();
         }
     }
@@ -663,15 +776,29 @@ impl BatchScanner {
             .map(|runner| runner.collect_results(block_tag, txid))
             .unwrap_or_default();
 
+        let orchard = self
+            .orchard_runner
+            .as_mut()
+            .map(|runner| runner.collect_results(block_tag, txid))
+            .unwrap_or_default();
+
         // Update the size of the batch scanner.
         metrics::decrement_gauge!(METRIC_SIZE_TXS, 1.0);
 
-        Box::new(BatchResult { sapling })
+        Box::new(BatchResult {
+            sapling,
+            orchard: OrchardDecryptedOutputs { inner: orchard },
+        })
     }
 }
 
 pub(crate) struct BatchResult {
     sapling: HashMap<(TxId, usize), DecryptedNote<[u8; 32], SaplingDomain>>,
+    orchard: OrchardDecryptedOutputs,
+}
+
+pub(crate) struct OrchardDecryptedOutputs {
+    pub(crate) inner: HashMap<(TxId, usize), DecryptedNote<[u8; 64], OrchardDomain>>,
 }
 
 impl BatchResult {
@@ -690,5 +817,9 @@ impl BatchResult {
                 },
             )
             .collect()
+    }
+
+    pub(crate) fn get_orchard(&self) -> &OrchardDecryptedOutputs {
+        &self.orchard
     }
 }